rename #43 : store 파일 이름 변경
