+++ src/App.tsx
@@ -6,7 +6,7 @@ import { Test } from './page/Test';
 import { MyAccount } from './page/auth/MyAccount';
 import { Home } from './page/home/Home';
 import { useAuthStore } from './stores/useAuthStore';
-import { useScreenConfigStore } from './stores/useDisplayConfigStore';
+import { useScreenConfigStore } from './stores/useScreenConfigStore';
 import { getAccessToken, getUserInfo } from './utils/localStorage';
 import elephantImg from '/elephant-bg.png';
